PoC/requirements.txt
pytlv
scipy
pycryptodome>=3.9.3
pyscard>=1.9.9